# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API with Go, MongoDB, and Swagger",
        "title": "Initial Project API",
        "contact": {
            "email": "fahim@example.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/carousels": {
            "get": {
                "description": "Retrieve all carousel items",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carousels"
                ],
                "summary": "Get all carousels",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/carousel.Carousel"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new carousel item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carousels"
                ],
                "summary": "Create a new carousel",
                "parameters": [
                    {
                        "description": "Carousel object",
                        "name": "carousel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/carousel.CarouselInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/carousel.Carousel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/carousels/{id}": {
            "put": {
                "description": "Update an existing carousel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carousels"
                ],
                "summary": "Update carousel by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Carousel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Carousel object",
                        "name": "carousel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/carousel.Carousel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a carousel item",
                "tags": [
                    "carousels"
                ],
                "summary": "Delete carousel by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Carousel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "API health status",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "carousel.Carousel": {
            "type": "object",
            "properties": {
                "Slug": {
                    "type": "string"
                },
                "createAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "updatedAt": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "carousel.CarouselInput": {
            "type": "object",
            "properties": {
                "isActive": {
                    "type": "boolean"
                },
                "slug": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    }
}